% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundedCovarianceDP.R
\name{boundedCovarianceDP}
\alias{boundedCovarianceDP}
\title{Differentially private covariance}
\usage{
boundedCovarianceDP(x, y, epsilon, x_min, x_max, y_min, y_max)
}
\arguments{
\item{x}{First input to the covariance}

\item{y}{Second input to the covariance}

\item{epsilon}{Privacy budget}

\item{x_min}{Lower bound for x}

\item{x_max}{Upper bound for x}

\item{y_min}{Lower bound for y}

\item{y_max}{Upper bound for y}
}
\value{
a numeric, differentially private covariance of the input vectors x and y
}
\description{
Compute a differential implementation of the covariance
}
