% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundedMeanDP.R
\name{boundedMeanDP}
\alias{boundedMeanDP}
\title{Differentially private mean}
\usage{
boundedMeanDP(input_data, epsilon, lower_bound, upper_bound)
}
\arguments{
\item{input_data}{the input vector}

\item{epsilon}{privacy budget}

\item{lower_bound}{lower bound for input values}

\item{upper_bound}{upper bound for input values}
}
\value{
a numeric, differentially private mean of the input vector
}
\description{
Computes a differentially private mean of a vector
}
