% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundedVarianceDP.R
\name{boundedVarianceDP}
\alias{boundedVarianceDP}
\title{Differentially private variance}
\usage{
boundedVarianceDP(input_data, epsilon, lower_bound, upper_bound)
}
\arguments{
\item{input_data}{the input vector}

\item{epsilon}{privacy budget}

\item{lower_bound}{lower bound for input values}

\item{upper_bound}{upper bound for input values}
}
\value{
a numeric, differentially private variance of the input vector
}
\description{
Computes a differentially private variance of a vector
}
