% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundedStandardDeviationDP.R
\name{boundedStandardDeviationDP}
\alias{boundedStandardDeviationDP}
\title{Title}
\usage{
boundedStandardDeviationDP(input_data, epsilon, lower_bound, upper_bound)
}
\arguments{
\item{upper_bound}{}
}
\description{
Title
}
